requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
apache-beam==2.61.0
asn1crypto==1.5.1
attrs==24.3.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cloud-sql-python-connector==1.15.0
cloudpickle==2.2.1
colorama==0.4.6
crcmod==1.7
cryptography==44.0.0
Deprecated==1.2.15
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
fastavro==1.10.0
fasteners==0.19
frozenlist==1.5.0
google-api-core==2.24.0
google-apitools==0.5.31
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-bigtable==2.27.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.2
google-cloud-dlp==3.26.0
google-cloud-language==2.16.0
google-cloud-pubsub==2.27.1
google-cloud-pubsublite==1.11.1
google-cloud-recommendations-ai==0.10.15
google-cloud-resource-manager==1.14.0
google-cloud-secret-manager==2.22.0
google-cloud-spanner==3.51.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.15.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
keyrings.google-artifactregistry-auth==1.1.2
more-itertools==10.5.0
multidict==6.1.0
numpy==2.1.3
oauth2client==4.1.3
objsize==0.7.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pg8000==1.31.2
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
scramp==1.4.5
shapely==2.0.6
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
